Fixed bugs and added List to select drivers for manager
Updated Sign up logic to add userId in firestore as well ,since key of USERS collection is email not uid. So to match that up with locations userId and get correct locations for correct users, had to fix that. Also implemented logic for managers to select specific drivers to view their locations. Still need to work on the UI better and fix some more bugs for manger