Update on Managers Dashboard
Updated the managers screen, so they can now see live users(drivers) who are currently signed in and tracking their location. Changed the drop down menu to have the current orders made on a list and can only be viewed by the managers. Through the list you can view the routes of orders.