Dashboard Update
Made a DashboardStyles.js to implement in the styles and connected it to Dashboard.js. This will make the Dashboard file shorter in code. Also fixed a few bugs on the Dashboard. Now when clicking on current orders you can see where the drivers left off, If they log out or their location isn't being tracked. This is only available for the manager role.